+++ eventlog-live/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The version 0.7.0.0 was skipped to avoid confusion with `eventlog-live-otelcol-0.7.0.0`.
 
-- Merge the executable `eventlog-live-otelcol` into this package.
+- Merge the executable `eventlog-live-otelcol` into this package as `eventlog-live-otlp`.
 
 ### 0.6.0.0